+++ src/pages/config/settings/components/snapshot/ScrapeUIDialog.tsx
