+++ src/pages/config/settings/components/snapshot/viewer/components/AccessTable.tsx
