+++ src/pages/config/settings/components/snapshot/viewer/components/AliasList.tsx
