+++ src/pages/config/settings/components/snapshot/viewer/components/ConfigNode.tsx
