+++ src/pages/config/settings/components/snapshot/viewer/components/DetailPanel.tsx
