+++ src/pages/config/settings/components/snapshot/viewer/components/EntityTable.tsx
