+++ src/pages/config/settings/components/snapshot/viewer/components/HARPanel.tsx
