+++ src/pages/config/settings/components/snapshot/viewer/components/ScrapeConfigPanel.tsx
