+++ src/pages/config/settings/components/snapshot/viewer/components/SnapshotPanel.tsx
