+++ src/pages/config/settings/components/snapshot/viewer/components/SplitPane.tsx
