+++ src/pages/config/settings/components/snapshot/viewer/globals.d.ts
